requirements.txt
pyttsx3
requests
paramiko
SpeechRecognition
wikipedia
wolframalpha
googlesearch-python
selenium
pyautogui
pyaudio
